+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Features
+
+- Show output file name for failing tests when using -o option
+
 ## [1.9.0] 2025-11-14
 
 ### Features